refactor to make owner and tag relationship mandatory.
The initial idea was that one would order tags and then once they
physically received them they would scan it an register it to
themselves. The idea was to sendout high quality tags and that you could
order them as gifts for other people. Now I pivoted to a slighly
different idea that requires that each tag is always associated to an
owner.